+++ R/meta-chk.R
@@ -195,7 +195,7 @@ chknull <- function(x, name = NULL) {
   invisible(NULL)
 }
 chknumeric <- function(x, min, max, zero = FALSE, length = 0,
-                       name = NULL, single = FALSE) {
+                       name = NULL, single = FALSE, integer = FALSE) {
   if (!missing(single) && single)
     length <- 1
   ##
@@ -238,6 +238,15 @@ chknumeric <- function(x, min, max, zero = FALSE, length = 0,
       (any(x < min, na.rm = TRUE) | any(x > max, na.rm = TRUE)))
     stop("Argument '", name, "' must be between ",
          min, " and ", max, ".", call. = FALSE)
+  ##
+  if (integer && any(!is.wholenumber(x))) {
+    if (length(x) == 1)
+      stop("Argument '", name, "' must be an integer.",
+           call. = FALSE)
+    else
+      stop("Argument '", name, "' may only contain integers.",
+           call. = FALSE)
+  }
   ##
   invisible(NULL)
 }