+++ R/settings.meta.R
@@ -83,7 +83,8 @@
 #'   tau-squared \cr
 #' \code{CIbracket}, \tab "[" \tab \cr
 #' \code{CIseparator} \tab ", " \tab print confidence intervals as
-#'   "\code{[., .]}"
+#'   "\code{[., .]}" \cr
+#' \code{header.line}, \tab TRUE \tab print header line
 #' }
 #'
 #' JAMA settings:
@@ -105,7 +106,8 @@
 #' \code{zero.pval}, \tab TRUE \tab print p-values with leading zero
 #' \cr
 #' \code{JAMA.pval}, \tab TRUE \tab round p-values to three digits
-#'   (for 0.001 < p \eqn{\le} 0.01) or two digits (p > 0.01)
+#'   (for 0.001 < p \eqn{\le} 0.01) or two digits (p > 0.01) \cr
+#' \code{header.line}, \tab TRUE \tab print header line
 #' }
 #' 
 #' IQWiG, General Methods 5 settings:
@@ -584,6 +586,8 @@ settings.meta <- function(..., quietly = TRUE) {
     setOption("forest.stat", TRUE)
     setOption("forest.Q.subgroup", TRUE)
     ##
+    setOption("header.line", FALSE)
+    ##
     setOption("fontsize", 12)
     setOption("fontfamily", NULL)
     setOption("fs.common", NULL)
@@ -652,7 +656,8 @@ settings.meta <- function(..., quietly = TRUE) {
                  "test.subgroup", "test.effect.subgroup",
                  "digits.I2", "digits.tau2", "digits.tau",
                  "CIbracket", "CIseparator",
-                 "zero.pval", "JAMA.pval"),
+                 "zero.pval", "JAMA.pval",
+                 "text.common", "text.w.common", "header.line"),
         new = list(replaceNULL(args[["method.random.ci"]], "classic"),
                    replaceNULL(args[["method.tau"]], "DL"),
                    replaceNULL(args[["tau.common"]], FALSE),
@@ -670,7 +675,10 @@ settings.meta <- function(..., quietly = TRUE) {
                    replaceNULL(args[["CIbracket"]], "["),
                    replaceNULL(args[["CIseparator"]], ", "),
                    replaceNULL(args[["zero.pval"]], FALSE),
-                   replaceNULL(args[["JAMA.pval"]], FALSE)
+                   replaceNULL(args[["JAMA.pval"]], FALSE),
+                   replaceNULL(args[["text.common"]], "Fixed effect model"),
+                   replaceNULL(args[["text.w.common"]], "fixed"),
+                   replaceNULL(args[["header.line"]], TRUE)
                    ),
         setting = "RevMan 5 settings",
         quietly = quietly)
@@ -681,19 +689,28 @@ settings.meta <- function(..., quietly = TRUE) {
                                 "test.subgroup", "test.effect.subgroup",
                                 "digits.I2", "digits.pval",
                                 "CIbracket", "CIseparator",
-                                "zero.pval", "JAMA.pval"),
-                       new = list("JAMA", TRUE,
-                                  FALSE, FALSE,
-                                  0, 3,
-                                  "(", "-",
-                                  FALSE, TRUE),
+                                "zero.pval", "JAMA.pval",
+                                "header.line"),
+                       new = list("JAMA",
+                                  replaceNULL(args[["test.overall"]], TRUE),
+                                  replaceNULL(args[["test.subgroup"]], FALSE),
+                                  replaceNULL(args[["test.effect.subgroup"]], FALSE),
+                                  replaceNULL(args[["digits.I2"]], 0),
+                                  replaceNULL(args[["digits.pval"]], 3),
+                                  replaceNULL(args[["CIbracket"]], "("),
+                                  replaceNULL(args[["CIseparator"]], "-"),
+                                  replaceNULL(args[["zero.pval"]], FALSE),
+                                  replaceNULL(args[["JAMA.pval"]], TRUE),
+                                  replaceNULL(args[["header.line"]], TRUE)),
                        setting = "JAMA settings",
                        quietly = quietly)
     }
     ##
     else if (setting == "IQWiG5") {
       specificSettings(args = c("method.random.ci", "prediction"),
-                       new = list("HK", TRUE),
+                       new = list(replaceNULL(args[["method.random.ci"]],
+                                              "HK"),
+                                  replaceNULL(args[["prediction"]], TRUE)),
                        setting = "IQWiG 5 settings",
                        quietly = quietly)
     }
@@ -701,7 +718,12 @@ settings.meta <- function(..., quietly = TRUE) {
     else if (setting == "IQWiG6") {
       specificSettings(args = c("method.random.ci", "adhoc.hakn.ci",
                                 "method.tau", "prediction"),
-                       new = list("HK", "IQWiG6", "PM", TRUE),
+                       new = list(replaceNULL(args[["method.random.ci"]],
+                                              "HK"),
+                                  replaceNULL(args[["adhoc.hakn.ci"]],
+                                              "IQWiG6"),
+                                  replaceNULL(args[["method.tau"]], "PM"),
+                                  replaceNULL(args[["prediction"]], TRUE)),
                        setting = "IQWiG 6 settings",
                        quietly = quietly)
     }
@@ -710,9 +732,14 @@ settings.meta <- function(..., quietly = TRUE) {
       specificSettings(args = c("method.tau", "exact.smd",
                                 "text.common", "text.w.common",
                                 "warn.deprecated"),
-                       new = list("DL", FALSE,
-                                  "Fixed effect model", "fixed",
-                                  FALSE),
+                       new = list(replaceNULL(args[["method.tau"]], "DL"),
+                                  replaceNULL(args[["exact.smd"]], FALSE),
+                                  replaceNULL(args[["text.common"]],
+                                              "Fixed effect model"),
+                                  replaceNULL(args[["text.w.common"]],
+                                              "fixed"),
+                                  replaceNULL(args[["warn.deprecated"]],
+                                              FALSE)),
                        setting =
                          "settings from meta, version 4 or below",
                        quietly = quietly)
@@ -720,7 +747,8 @@ settings.meta <- function(..., quietly = TRUE) {
     ##
     else if (setting == "geneexpr") {
       specificSettings(args = c("scientific.pval", "method.tau.ci"),
-                       new = list(TRUE, ""),
+                       new = list(replaceNULL(args[["scientific.pval"]], TRUE),
+                                  replaceNULL(args[["method.tau.ci"]], "")),
                        setting = "Settings for gene expression data",
                        quietly = quietly)
     }
@@ -919,6 +947,8 @@ settings.meta <- function(..., quietly = TRUE) {
     ##
     catarg("forest.stat            ")
     catarg("forest.Q.subgroup      ")
+    ##
+    catarg("header.line            ")
     ##
     catarg("fontsize               ")
     catarg("fontfamily             ")
@@ -1208,6 +1238,8 @@ settings.meta <- function(..., quietly = TRUE) {
     ##
     setlogical("forest.stat", args)
     setlogical("forest.Q.subgroup", args)
+    ##
+    setlogical("header.line", args)
     ##
     setnumeric("fontsize", args)
     setcharacter("fontfamily", args, NULL.ok = TRUE)